Commitan ke 4 bro

+++ src/main/java/com/jaxi/repository/AreaSalesManagerRepository.java
@@ -3,5 +3,5 @@ package com.jaxi.repository;
 import com.jaxi.entity.AreaSalesManager;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AreaSalesManagerRepository extends JpaRepository<AreaSalesManager,Integer> {
+public interface AreaSalesManagerRepository extends JpaRepository<AreaSalesManager,Long> {
 }
