+++ src/main/java/com/jaxi/repository/TeamLeaderRepository.java
@@ -3,6 +3,6 @@ package com.jaxi.repository;
 import com.jaxi.entity.TeamLeader;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TeamLeaderRepository  extends JpaRepository<TeamLeader, Integer> {
+public interface TeamLeaderRepository  extends JpaRepository<TeamLeader, Long> {
 
 }
